--- src/main/java/roomescape/Controller/ReservationTimeController.java
@@ -1,39 +1,0 @@
-package roomescape.controller;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-import roomescape.dto.ReservationTimeRequest;
-import roomescape.dto.ReservationTimeResponse;
-import roomescape.service.ReservationTimeService;
-
-import java.util.List;
-
-@RestController
-@RequestMapping(path = "/times")
-public class ReservationTimeController {
-    private final ReservationTimeService reservationTimeService;
-
-    public ReservationTimeController(ReservationTimeService reservationTimeService) {
-        this.reservationTimeService = reservationTimeService;
-    }
-
-    @GetMapping()
-    public ResponseEntity<List<ReservationTimeResponse>> read() {
-        List<ReservationTimeResponse> reservationTimes = reservationTimeService.findAll();
-        return ResponseEntity.status(HttpStatus.OK).body(reservationTimes);
-    }
-
-    @PostMapping()
-    public ResponseEntity<ReservationTimeResponse> create(@RequestBody ReservationTimeRequest request) {
-        long id = reservationTimeService.add(request);
-        ReservationTimeResponse reservationTime = reservationTimeService.findOne(id);
-        return ResponseEntity.status(HttpStatus.CREATED).body(reservationTime);
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> delete(@PathVariable("id") Long id) {
-        reservationTimeService.delete(id);
-        return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
-    }
-}
