--- src/main/java/roomescape/Controller/ThemeController.java
@@ -1,39 +1,0 @@
-package roomescape.controller;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-import roomescape.dto.ThemeRequest;
-import roomescape.dto.ThemeResponse;
-import roomescape.service.ThemeService;
-
-import java.util.List;
-
-@RestController
-@RequestMapping("/themes")
-public class ThemeController {
-    private final ThemeService themeService;
-
-    public ThemeController(ThemeService themeService) {
-        this.themeService = themeService;
-    }
-
-    @GetMapping()
-    public ResponseEntity<List<ThemeResponse>> read() {
-        List<ThemeResponse> themes = themeService.findAll();
-        return ResponseEntity.status(HttpStatus.OK).body(themes);
-    }
-
-    @PostMapping()
-    public ResponseEntity<ThemeResponse> create(@RequestBody ThemeRequest request) {
-        long themeId = themeService.add(request);
-        ThemeResponse theme = themeService.findOne(themeId);
-        return ResponseEntity.status(HttpStatus.CREATED).body(theme);
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> delete(@PathVariable Long id) {
-        themeService.delete(id);
-        return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
-    }
-}
