--- src/main/java/roomescape/DTO/ExceptionResponse.java
@@ -1,17 +1,0 @@
-package roomescape.dto;
-
-public class ExceptionResponse {
-    private final String message;
-
-    private ExceptionResponse(String message) {
-        this.message = message;
-    }
-
-    public static ExceptionResponse createByMessage(String message) {
-        return new ExceptionResponse(message);
-    }
-
-    public String getMessage() {
-        return message;
-    }
-}
