--- src/main/java/roomescape/DTO/ReservationTimeRequest.java
@@ -1,20 +1,0 @@
-package roomescape.dto;
-
-public class ReservationTimeRequest {
-    private String startAt;
-
-    private ReservationTimeRequest() {
-    }
-
-    private ReservationTimeRequest(String startAt) {
-        this.startAt = startAt;
-    }
-
-    public static ReservationTimeRequest create(String startAt) {
-        return new ReservationTimeRequest(startAt);
-    }
-
-    public String getStartAt() {
-        return this.startAt;
-    }
-}
