--- src/main/java/roomescape/DTO/ThemeResponse.java
@@ -1,48 +1,0 @@
-package roomescape.dto;
-
-import roomescape.entity.Theme;
-
-import java.util.List;
-
-public class ThemeResponse {
-    private final Long id;
-    private final String name;
-    private final String description;
-    private final String thumbnail;
-
-    private ThemeResponse(Long id, String name, String description, String thumbnail) {
-        this.id = id;
-        this.name = name;
-        this.description = description;
-        this.thumbnail = thumbnail;
-    }
-
-    public static ThemeResponse from(Theme theme) {
-        return new ThemeResponse(
-                theme.getId(),
-                theme.getName(),
-                theme.getDescription(),
-                theme.getThumbnail()
-        );
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public String getThumbnail() {
-        return thumbnail;
-    }
-
-    public static List<ThemeResponse> toResponses(List<Theme> themes) {
-        return themes.stream().map(ThemeResponse::from).toList();
-    }
-}
