--- src/main/java/roomescape/Repository/ReservationTimeRepository.java
@@ -1,14 +1,0 @@
-package roomescape.repository;
-
-import roomescape.entity.ReservationTime;
-
-import java.util.List;
-
-public interface ReservationTimeRepository {
-    List<ReservationTime> findAll();
-    ReservationTime findById(Long id);
-    ReservationTime findByStartAt(String startAt);
-    Long countReservationMatchWith(Long id);
-    long save(String startAt);
-    long cancelById(Long id);
-}
